Update to checkCommand DG
