+++ docs/diagrams/CheckSequenceDiagram.puml
@@ -1,0 +1,69 @@
+@startuml
+!include style.puml
+
+box Logic LOGIC_COLOR_T1
+participant ":LogicManager" as LogicManager LOGIC_COLOR
+participant ":FinSecParser" as FinSecParser LOGIC_COLOR
+participant ":CheckCommandParser" as CheckCommandParser LOGIC_COLOR
+participant "check:CheckCommand" as CheckCommand LOGIC_COLOR
+participant ":CommandResult" as CommandResult LOGIC_COLOR
+end box
+
+box Model MODEL_COLOR_T1
+participant ":Model" as Model MODEL_COLOR
+end box
+
+[-> LogicManager : execute("check 1")
+activate LogicManager
+
+LogicManager -> FinSecParser : parseCommand("check 1")
+activate FinSecParser
+
+create CheckCommandParser
+FinSecParser -> CheckCommandParser
+activate CheckCommandParser
+
+CheckCommandParser --> FinSecParser
+deactivate CheckCommandParser
+
+FinSecParser -> CheckCommandParser : parse("1")
+activate CheckCommandParser
+
+create CheckCommand
+CheckCommandParser -> CheckCommand
+activate CheckCommand
+
+CheckCommand --> CheckCommandParser : check
+deactivate CheckCommand
+
+CheckCommandParser --> FinSecParser : check
+deactivate CheckCommandParser
+'Hidden arrow to position the destroy marker below the end of the activation bar.
+CheckCommandParser -[hidden]-> FinSecParser
+destroy CheckCommandParser
+
+FinSecParser --> LogicManager : check
+deactivate FinSecParser
+
+LogicManager -> CheckCommand : execute(String commandText)
+activate CheckCommand
+
+CheckCommand -> Model : handleClaim(Claim claim)
+activate Model
+
+Model --> CheckCommand
+deactivate Model
+
+create CommandResult
+CheckCommand -> CommandResult
+activate CommandResult
+
+CommandResult --> CheckCommand
+deactivate CommandResult
+
+CheckCommand --> LogicManager : commandResult
+deactivate GotoCommand
+
+[<--LogicManager
+deactivate LogicManager
+@enduml
