+++ src/main/java/seedu/address/model/ModelManager.java
@@ -9,14 +9,12 @@ import java.util.logging.Logger;
 
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
-import javafx.fxml.FXML;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.model.autocorrectsuggestion.AutocorrectSuggestion;
 import seedu.address.model.claim.Claim;
 import seedu.address.model.contact.Contact;
 import seedu.address.model.income.Income;
-import seedu.address.ui.IndividualClaimWindow;
 import seedu.address.ui.UiManager;
 
 /**
@@ -168,20 +166,6 @@ public class ModelManager implements Model {
     public boolean hasContactFor(Claim claim) {
         requireNonNull(claim);
         return finSec.hasContact(claim.getName(), claim.getPhone());
-    }
-
-    /**
-     * Opens the claim window or focuses on it if it's already opened.
-     */
-    @FXML
-    public static void handleClaim(Claim claim) {
-        IndividualClaimWindow individualClaimWindow = new IndividualClaimWindow(claim);
-
-        if (!individualClaimWindow.isShowing()) {
-            individualClaimWindow.show();
-        } else {
-            individualClaimWindow.focus();
-        }
     }
 
     //=========== Suggestions ================================================================================
